refactor(home): 스타일 관련 코드 제거
