+++ apps/home/src/pages/home/constants/sections.ts
@@ -4,7 +4,7 @@ export const HOME_SECTIONS: Section[] = [
   {
     id: "blog",
     label: "Blog",
-    description: "생각과 경험을 기록합니다",
+    description: "경험을 사유하고 기록합니다",
     href: "/blog",
     color: "from-blue-500 to-cyan-500",
     size: "lg",
@@ -13,7 +13,7 @@ export const HOME_SECTIONS: Section[] = [
   {
     id: "portfolio",
     label: "Portfolio",
-    description: "만든 것들을 모았습니다",
+    description: "창작물을 모았습니다",
     href: "/portfolio",
     color: "from-purple-500 to-pink-500",
     size: "md",
@@ -21,7 +21,7 @@ export const HOME_SECTIONS: Section[] = [
   {
     id: "resume",
     label: "Resume",
-    description: "경력과 스킬을 정리했습니다",
+    description: "이력을 정리했습니다",
     href: "/resume",
     color: "from-orange-500 to-red-500",
     size: "sm",
